ttymsg: fix compiler warnings and use EXIT_
[kzak@redhat.com: - remove 'register' storage class specifier]

+++ term-utils/ttymsg.h
@@ -1,2 +1,2 @@
-char *ttymsg(struct iovec *iov, int iovcnt, char *line, int tmout);
+char *ttymsg(struct iovec *iov, size_t iovcnt, char *line, int tmout);
